Fix last conflict between CRUD_USER and main. Add menu bar User CRUD. Add Admin access on CREATE UPDATE DELETE User

+++ src/main/java/com/filmplanner/dao/UserDAO.java
@@ -8,18 +8,18 @@ import java.util.List;
 public interface UserDAO {
     User findByEmail(String email);
 
-    long update(int id, User updatedUser);
+    int update(Long id, User updatedUser);
 
     List<User> findAll();
 
-    User findById(int id);
+    User findById(Long id);
 
     /**
      * @param newUser
      * @return 1 if successful, 0 if not
      * creates user in the database
      */
-    int create(User newUser);
+    Long create(User newUser);
 
-    int deleteById(int id);
+    int deleteById(Long id);
 }
